Perform full check on mempool in IsCertApplicableToState.
IsCertApplicableToState is updated so that whenever we add a certificate to the mempool, we perform a full check to
assess the correctness of the ordering of all the certificates in mempool for the specific sidechain.
The previous implementation relied solely on a maxReferencedHeight sc-specific variable.

Also, this commit adds a new testCheckReferencedHeight gtest, as a unit test for the function above.