Fixed a sporadic issue on sc_cert_ceasing.py
The test includes a restart of nodes triggered immediately after a FT transaction is added to the wallet of one node.

Due to a race condition, the stop of the nodes could happen after or before the transaction was actually stored in the wallet. If it happened before, everything worked smoothly, otherwise, after the restart, the first node ended up having that transaction in the mempool without relaying it to the other peer, causing the sync_all() function to hang up.

Now the startup parameter `-zapwallettxes=2` has been added to remove that transaction from the wallet and avoid any misalignment.