Added certificate dependency on json returned by getblocktemplate
Field 'depends' in the json returned by getblocktemplate now includes dependencies
between tx and certificates (certs that spend vouts of txs which are currently in mempool).
We also added a 'certdepends' in the same json which includes dependencies among
certificates (certs which spend vouts of other certificates in the mempool).